% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSTN_SAS.R
\name{BSTN_SAS}
\alias{BSTN_SAS}
\title{Fit Bayesian Skewed Tensor Normal (BSTN) model with tensor spike-and-slab prior to GAAD data
Lee et al. (2021+) Bayesian Regression Analysis of Skewed Tensor Response}
\usage{
BSTN_SAS(Y, X, vecy, n.burn = 10, n.save = 100, thin = 1)
}
\arguments{
\item{Y}{t x s x b x n array of skewed tensor response}

\item{X}{p x n matrix of covariates}

\item{vecy}{tsb x n matrix: vectorized tensor}

\item{n.burn}{burn-in period}

\item{n.save}{number of posterior samples}

\item{thin}{thinning size}
}
\value{
Returns a list with the following components:
\item rho: posterior samples of correlation of each mode of tensor response (3 x n.save)
\item sigma.sq: posterior samples of variance parameter (b x n.save)
\item lam.est: posterior samples of skewness parameters (b x n.save)
\item est.est: posterior samples of common effects of covariates (p x n.save)
\item omega: posterior samples of (zeros/ones) that particular element is
             included in the gamma.est (t x s x b x p x n.save)
\item gamma.est: posterior samples of sparsity elements that
                 identify different effects of each tooth-sites (t x s x b x p x n.save)
}
\description{
Fit Bayesian Skewed Tensor Normal (BSTN) model with tensor spike-and-slab prior to GAAD data
Lee et al. (2021+) Bayesian Regression Analysis of Skewed Tensor Response
}
